+++ armyknife/webexbothandler.py
@@ -24,7 +24,7 @@ class WebexTeamsBotHandler:
         if not self.spark.is_actor_user:
             self.spark.link.post_message(
                 self.spark.room_id,
-                "**Welcome to Cisco Webex Teams Army Knife, " + self.spark.person_object +
+                "**Welcome to Army Knife Army Knife, " + self.spark.person_object +
                 "!**\n\n Please authorize the app by clicking the following link: " +
                 url + "/www",
                 markdown=True)
@@ -44,7 +44,7 @@ class WebexTeamsBotHandler:
             if self.spark.room_type == 'group':
                 self.spark.link.post_bot_message(
                     spark_id=self.spark.room_id,
-                    text="**Welcome to Cisco Webex Teams Army Knife!**\n\n To use, please create a 1:1 room with " +
+                    text="**Welcome to Army Knife Army Knife!**\n\n To use, please create a 1:1 room with " +
                          self.spark.config.bot['email'] +
                          ". If you don't get an answer, type /init in that room.",
                     markdown=True)
@@ -228,7 +228,7 @@ class WebexTeamsBotHandler:
             self.spark.link.post_admin_message(out, markdown=True)
         elif self.spark.cmd == "/help":
             self.spark.link.post_admin_message(
-                "**Cisco Webex Teams Army Knife: Admin Help**\n\n"
+                "**Army Knife: Admin Help**\n\n"
                 "Use `/mail <email> message` to send somebody a message from the bot.\n\n"
                 "Use `/all-users` for listing and messaging all users.\n\n"
                 "Use `/stats` to see statistics on command usage.",
@@ -243,38 +243,38 @@ class WebexTeamsBotHandler:
         if self.spark.cmd == '/help' or ('/' not in self.spark.cmd and len(self.spark.msg_list_wcap) <= 2):
             self.spark.link.post_bot_message(
                 spark_id=self.spark.room_id,
-                text="**Hi there from the Cisco Webex Teams Army Knife!**\n\n"
+                text="**Hi there from the Army Knife!**\n\n"
                      "To use, please create a 1:1 room with the bot (" +
                      self.spark.config.bot['email'] + ").",
                 markdown=True)
             if not self.spark.is_actor_user:
                 self.spark.link.post_bot_message(
                     email=self.spark.person_object,
-                    text="**Hi there from the Cisco Webex Teams Army Knife!**\n\n"
+                    text="**Hi there from the Army Knife!**\n\n"
                          "Please type /init in the 1:1 room to authorize the app.",
                     markdown=True)
 
     def help(self):
         self.spark.link.post_bot_message(
             email=self.spark.person_object,
-            text="**Cisco Webex Teams Army Knife (author: Greger Wedel)**\n\n"
+            text="**Army Knife (author: Greger Wedel)**\n\n"
                  "Help message for commands that only work in the bot 1:1 room.\n\n"
                  "**App Management**\n\n"
                  "- Use `/init` to authorize the app so that all commands work.\n\n"
-                 "- Use `/delete DELETENOW` to delete your Cisco Webex Teams Army Knife account, this room, and "
+                 "- Use `/delete DELETENOW` to delete your Army Knife account, this room, and "
                  "all data associated with this account.\n\n"
-                 "- Use `/disable` to temporarily disable the Cisco Webex Teams Army Knife account.\n\n"
-                 "- Use `/enable` to enable the Cisco Webex Teams Army Knife account.\n\n"
+                 "- Use `/disable` to temporarily disable the Army Knife account.\n\n"
+                 "- Use `/enable` to enable the Army Knife account.\n\n"
                  "- Use `/support <message>` to send a message to support.\n\n"
-                 "- Use `/me` to get info about your Cisco Webex Teams Army Knife account.\n\n"
+                 "- Use `/me` to get info about your Army Knife account.\n\n"
                  "- Use `/recommend <email> <message>` to send a message to another user and recommend "
-                 "Cisco Webex Teams Army Knife.\n\n"
+                 "Army Knife.\n\n"
                  "- Use `/nomentionalert` to turn off 1:1 bot room alerts on mentions.\n\n"
                  "- Use `/mentionalert` to turn on (default) 1:1 bot room alerts on mentions.\n\n"
                  "- Use `/noroomalert` to turn off 1:1 bot room alerts on new rooms.\n\n"
                  "- Use `/roomalert` to turn on (default) 1:1 bot room alerts on new rooms.\n\n"
-                 "- Use `/noannouncements` to turn off announcements about Cisco Webex Teams Army Knife.\n\n"
-                 "- Use `/announcements` to turn on (default) announcements about Cisco Webex Teams Army Knife.\n\n"
+                 "- Use `/noannouncements` to turn off announcements about Army Knife.\n\n"
+                 "- Use `/announcements` to turn on (default) announcements about Army Knife.\n\n"
                  "**Todo/Followup List (aliases /followup and /fu)**\n\n"
                  "- Use `/todo Todo item...` to set a new todo item.\n\n"
                  "- Use `/todo` to list your todo items.\n\n"
@@ -290,7 +290,7 @@ class WebexTeamsBotHandler:
                  "- Use `/topofmind reminder on|off` to set or stop a daily reminder of your list at "
                  "this time.\n\n"
                  "**Box.com Integration**\n\n"
-                 "- Use `/box <rootfolder>` to add a Box account to Cisco Webex Teams Army Knife. "
+                 "- Use `/box <rootfolder>` to add a Box account to Army Knife. "
                  "Optionally specify the folder where all Army Knife Box folders will be created.\n\n"
                  "- Use `/nobox` to disconnect and delete the Box service.\n\n"
                  "**Room management**\n\n"
@@ -300,6 +300,7 @@ class WebexTeamsBotHandler:
                  "- Use `/deletemember <email> <room-id,room-id...>` to delete a user from a room"
                  " or list of rooms "
                  "(use Cisco Webex Teams Id from e.g. /checkmember or /listroom).\n\n"
+                 "- Use `/deletemember <email> FORCE count` to delete a user from ALL shared rooms.\n\n"
                  "- Use `/addmember <email> <room-id,room-id...>` to add a user to a room or"
                  " list of rooms "
                  "(use Cisco Webex Teams Id from e.g. /checkmember or /listroom).\n\n"
@@ -503,7 +504,7 @@ class WebexTeamsBotHandler:
             topofmind['title'] = self.spark.msg_data['text'][
                                  len(self.spark.msg_list_wcap[0]) +
                                  len(self.spark.msg_list_wcap[1]) + 2:]
-            out = json.dumps(topofmind)
+            out = json.dumps(topofmind, ensure_ascii=False)
             self.spark.me.set_property('topofmind', out)
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
@@ -586,7 +587,7 @@ class WebexTeamsBotHandler:
                     text="Added list item **" + str(index) + "** with text `" + toplist[index] + "`",
                     markdown=True)
             topofmind['list'] = toplist
-            out = json.dumps(topofmind, sort_keys=True)
+            out = json.dumps(topofmind, sort_keys=True, ensure_ascii=False)
             self.spark.me.set_property('topofmind', out)
             self.spark.me.register_diffs(target='properties', subtarget='topofmind', blob=out)
             # List out the updated list
@@ -720,7 +721,7 @@ class WebexTeamsBotHandler:
                 text="Added list item **" + str(index+1) + "** with text `" + toplist[index] + "`",
                 markdown=True)
         todo['list'] = toplist
-        out = json.dumps(todo, sort_keys=True)
+        out = json.dumps(todo, sort_keys=True, ensure_ascii=False)
         self.spark.me.set_property('todo', out)
         # List out the updated list
         out = "**" + todo['title'] + "**"
@@ -811,7 +812,7 @@ class WebexTeamsBotHandler:
                 else:
                     self.spark.link.post_bot_message(
                         email=self.spark.person_object,
-                        text="Welcome to the new Cisco Webex Teams Army Knife!\n\n"
+                        text="Welcome to the new Army Knife!\n\n"
                              "Not able to re-initialize properly from old Army Knife."
                              "Your account has probably timed out. Please do /init (again)",
                         markdown=True)
@@ -837,7 +838,7 @@ class WebexTeamsBotHandler:
             self.spark.me.delete_property('app_disabled')
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text="The Cisco Webex Teams Army Knife has now been enabled and will "
+                text="The Army Knife has now been enabled and will "
                      "process messages and respond to commands.")
             return
         if not self.spark.me or not self.spark.me.id:
@@ -855,7 +856,7 @@ class WebexTeamsBotHandler:
             self.spark.me.set_property('app_disabled', 'true')
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text="The Cisco Webex Teams Army Knife has now been disabled and will"
+                text="The Army Knife has now been disabled and will"
                      " not respond or process commands until you do "
                      "/enable. \n\n/enable, /help, and /init are the only commands that still work.")
             return
@@ -873,7 +874,7 @@ class WebexTeamsBotHandler:
                 firehose = "<none>"
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text="**Your Cisco Webex Teams Army Knife Account**\n\n----\n\n" +
+                text="**Your Army Knife Account**\n\n----\n\n" +
                      "**Registered email**: " + self.spark.me.creator + "\n\n" +
                      "**URL**: " + self.spark.config.root + self.spark.me.id + '/www\n\n' +
                      "**Webhook**: " + firehose +
@@ -885,7 +886,7 @@ class WebexTeamsBotHandler:
                 props = self.spark.me.get_properties()
                 self.spark.link.post_bot_message(
                     email=self.spark.person_object,
-                    text="**Your Cisco Webex Teams Army Knife Account Data**\n\n----\n\n" +
+                    text="**Your Army Knife Account Data**\n\n----\n\n" +
                          json.dumps(props, sort_keys=True, indent=4),
                     markdown=True)
         elif self.spark.cmd == '/delete':
@@ -919,12 +920,12 @@ class WebexTeamsBotHandler:
                       len(self.spark.msg_list_wcap[1]) + 2:]
             self.spark.link.post_bot_message(
                 email=self.spark.msg_list[1],
-                text=message + "\n\n**This bot, Cisco Webex Teams Army Knife, was recommended to you by " +
+                text=message + "\n\n**This bot, Army Knife, was recommended to you by " +
                 self.spark.person_object + "**\n\nType /init to get started!",
                 markdown=True)
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text=self.spark.msg_list[1] + " was just invited to use Cisco Webex Teams Army Knife!",
+                text=self.spark.msg_list[1] + " was just invited to use Army Knife!",
                 markdown=True)
             self.spark.link.post_admin_message(
                 text=self.spark.msg_list[1] + " was just recommended Army Knife by " +
@@ -966,9 +967,9 @@ class WebexTeamsBotHandler:
             self.spark.me.set_property('no_announcements', 'true')
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text="You will no longer get Cisco Webex Teams Army Knife announcements.")
+                text="You will no longer get Army Knife announcements.")
         elif self.spark.cmd == '/announcements':
             self.spark.me.delete_property('no_announcements')
             self.spark.link.post_bot_message(
                 email=self.spark.person_object,
-                text="You will now get Cisco Webex Teams Army Knife announcements!")
+                text="You will now get Army Knife announcements!")